added frontend

+++ src/main/java/org/example/model/dto/response/AccountResponseDTO.java
@@ -11,11 +11,11 @@ import java.util.List;
 
 @Data
 public class AccountResponseDTO {
-  private Long moneyAmount;
+  private int moneyAmount;
   private String secretWord;
   private AccountStatus status;
   private LocalDateTime creationDate;
-  private Long depositLimit;
+  private int depositLimit;
   private Client client;
   private List<Card> card;
   private CurrencyType currencyType;
